project/Face-Emotion-Detection/requirements.txt
streamlit==1.29.0
opencv-python-headless==4.8.1.78
tensorflow==2.15.0
numpy==1.24.3
keras==2.15.0
Pillow==10.1.0
